requirements.txt
arize-phoenix==12.23.0
arize-phoenix-client==1.26.0
arize-phoenix-evals==2.7.0
arize-phoenix-otel==0.14.0
dspy==3.0.3
litellm==1.78.5
llvmlite==0.44.0
numpy==1.26.4
numpydoc==1.8.0
ollama==0.4.9
openai==2.5.0
openai-harmony==0.0.8
opencv-python-headless==4.11.0.86
openinference-instrumentation==0.1.42
openinference-instrumentation-dspy==0.1.31
openinference-instrumentation-litellm==0.1.28
openinference-semantic-conventions==0.1.25
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-exporter-otlp-proto-http==1.39.0
opentelemetry-instrumentation==0.48b0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
opentelemetry-semantic-conventions-ai==0.4.3
pandas==2.3.3
pandas-datareader==0.10.0
pandocfilters==1.5.1
phoenix==0.9.1
pyarrow==17.0.0
safetensors==0.4.5
scikit-learn==1.7.2
scipy==1.16.2
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
vllm==0.13.0

